Reworked base threading classes thereby fixing compilations problems on Windows
Using std::mutex leads to problems with /clr code on Windows, where this header is not allowed. So this requires wrapper classes to hide the std::mutex details. In our base threading library we already have such wrappers but they require glib. Since we want to lower usage of glib in order to be able to get rid of it one day, the code has been changed to use C++11 threading classes instead.

In this process the RecMutexTryLock was removed as it provided something very special used only in one place. Instead the loop lock attempts have been implemented as direct lock attempts on the RecMutex class.

Further changes included here:
- The symbol table class now has an own mutex to sync access to it. It's no longer necessary to pass around a separate mutex instance.
- base::init_threading() is gone now.
- Server ping (select 1) has been changed to do 1, which does the same except no result set is returned.
- Semaphore::tryWait is not used anywhere, so it has not been ported over. The 2 unit tests for it were removed as well.
- A few outdated references to the ANTL3 runtime have been removed.
- Some smaller cleanup.