Bug #25671681 VIEW AND FUNCTION CHANGES GET LOST WHEN SYNCHRONIZING MODEL
This patch fixes not only the mentioned bug but any potential problem when running a command from a menu while edits are still ongoing.

Additionally, a threading problem was solved for macOS and the changed file reformatted to our standards.

Finally, a code generation bug was fixed that created invalid SQL failing so the entire synchronization.