Bug #25645222 ALTERING A STORED PROCEDURE DROPS NEW NAME, NOT OLD
The problem here was that grt DB objects didn't track their old names, which is important for correct renames (e.g. in object editors). There were actually 2 problems here: the one reported in the bug (old name was set to new name, so the alter wizard tried to drop the new object, instead of the old one) and after the object was changed the new name wasn't set as the new old one (to support further renames). The fix for these also required to remove calls to set the old name when parsing objects (except for new ones), as that would overwrite the existing old name.

In addition to the fixes for these problem a workaround has been removed for old servers, where the wrong name casing was used on macOS under certain circumstances. I couldn't reproduce this behavior anymore with server 5.7 and 8.