Added error handling for reverse engineering
With the parser switch we lost error reporting when reverse engineering a script. All parse errors where silently ignored and the script import failed for no apparent reason. This patch fixes that, with these changes:

- The parser module now stores parse errors in a grt list provided by the caller.
- The statement splitter now tracks and reports the current line for statements which allows to print the exact error line number. For this an own StatementRange type has been introduced and is used for determineStatementRanges.
- Corrected unit tests for the changes.
- Some code cleanup was performed on the way (correct variable names and casts mostly).

Additionally, this patch fixes a bug which failed the rev-eng if ANSI quoted dumps where used.
- Added a new checkbox to the rev-eng wizard that allows to specify ANSI_QUOTES mode.
- The sql mode ANSI_QUOTES is now set for the parser when selected by the user.